refactor: mvoe all Python-related code to "python" submodules
