+++ src/types/excelreader/python.rs
@@ -1,144 +1,30 @@
-use std::{
-    fs::File,
-    io::{BufReader, Cursor},
-};
-
-#[cfg(feature = "pyarrow")]
 use arrow_array::RecordBatch;
-#[cfg(feature = "pyarrow")]
-use arrow_pyarrow::ToPyArrow;
-#[cfg(feature = "python")]
-use pyo3::{Bound, IntoPyObjectExt, PyAny, PyResult, Python, pyclass, pymethods, types::PyString};
+use pyo3::{Bound, IntoPyObjectExt, PyAny, PyResult, Python, pymethods, types::PyString};
 
-use calamine::{
-    Data, DataRef, HeaderRow, Range, Reader, ReaderRef, Sheet as CalamineSheet, Sheets, Table,
-    open_workbook_auto, open_workbook_auto_from_rs,
-};
-
-#[cfg(feature = "python")]
-use crate::{data::record_batch_from_data_and_columns, error::py_errors::IntoPyResult};
-
+use super::ExcelReader;
 use crate::{
-    data::ExcelSheetData,
-    error::{ErrorContext, FastExcelError, FastExcelErrorKind, FastExcelResult},
+    ExcelSheet, ExcelTable,
+    data::{ExcelSheetData, record_batch_from_data_and_columns},
+    error::{ErrorContext, FastExcelErrorKind, FastExcelResult, py_errors::IntoPyResult},
     types::{
         dtype::{DTypeCoercion, DTypes},
+        excelreader::LoadSheetOptions,
+        excelsheet::{
+            Header, Pagination, SelectedColumns,
+            column_info::{build_available_columns_info, finalize_column_info},
+        },
         idx_or_name::IdxOrName,
-        python::excelsheet::table::{extract_table_names, extract_table_range},
     },
     utils::schema::get_schema_sample_rows,
 };
 
-use super::excelsheet::{
-    ExcelSheet, Header, Pagination, SelectedColumns,
-    column_info::{build_available_columns_info, finalize_column_info},
-};
-use super::table::ExcelTable;
-
-enum ExcelSheets {
-    File(Sheets<BufReader<File>>),
-    Bytes(Sheets<Cursor<Vec<u8>>>),
-}
-
-impl ExcelSheets {
-    fn worksheet_range(&mut self, name: &str) -> FastExcelResult<Range<Data>> {
-        match self {
-            Self::File(sheets) => sheets.worksheet_range(name),
-            Self::Bytes(sheets) => sheets.worksheet_range(name),
-        }
-        .map_err(|err| FastExcelErrorKind::CalamineError(err).into())
-        .with_context(|| format!("Error while loading sheet {name}"))
-    }
-
-    #[allow(dead_code)]
-    fn sheet_metadata(&self) -> &[CalamineSheet] {
-        match self {
-            ExcelSheets::File(sheets) => sheets.sheets_metadata(),
-            ExcelSheets::Bytes(sheets) => sheets.sheets_metadata(),
-        }
-    }
-
-    fn table_names(&mut self, sheet_name: Option<&str>) -> FastExcelResult<Vec<String>> {
-        match self {
-            Self::File(sheets) => {
-                extract_table_names(sheets, sheet_name)?.map(|v| v.into_iter().cloned().collect())
-            }
-            Self::Bytes(sheets) => {
-                extract_table_names(sheets, sheet_name)?.map(|v| v.into_iter().cloned().collect())
-            }
-        }
-    }
-
-    fn supports_by_ref(&self) -> bool {
-        matches!(
-            self,
-            Self::File(Sheets::Xlsx(_)) | Self::Bytes(Sheets::Xlsx(_))
-        )
-    }
-
-    fn with_header_row(&mut self, header_row: HeaderRow) -> &mut Self {
-        match self {
-            Self::File(sheets) => {
-                sheets.with_header_row(header_row);
-                self
-            }
-            Self::Bytes(sheets) => {
-                sheets.with_header_row(header_row);
-                self
-            }
-        }
-    }
-
-    fn worksheet_range_ref(&mut self, name: &str) -> FastExcelResult<Range<DataRef<'_>>> {
-        match self {
-            ExcelSheets::File(Sheets::Xlsx(sheets)) => Ok(sheets.worksheet_range_ref(name)?),
-            ExcelSheets::Bytes(Sheets::Xlsx(sheets)) => Ok(sheets.worksheet_range_ref(name)?),
-            _ => Err(FastExcelErrorKind::Internal(
-                "sheets do not support worksheet_range_ref".to_string(),
-            )
-            .into()),
-        }
-        .with_context(|| format!("Error while loading sheet {name}"))
-    }
-
-    fn get_table(&mut self, name: &str) -> FastExcelResult<Table<Data>> {
-        match self {
-            Self::File(sheets) => extract_table_range(name, sheets)?,
-            Self::Bytes(sheets) => extract_table_range(name, sheets)?,
-        }
-    }
-}
-
-#[cfg_attr(feature = "python", pyclass(name = "_ExcelReader"))]
-pub struct ExcelReader {
-    sheets: ExcelSheets,
-    sheet_metadata: Vec<CalamineSheet>,
-    source: String,
-}
-
 impl ExcelReader {
-    #[cfg(feature = "python")]
     fn build_selected_columns(
         use_columns: Option<&Bound<'_, PyAny>>,
     ) -> FastExcelResult<SelectedColumns> {
         use_columns.try_into().with_context(|| format!("expected selected columns to be list[str] | list[int] | str | Callable[[ColumnInfo], bool] | None, got {use_columns:?}"))
     }
 
-    // NOTE: Not implementing TryFrom here, because we're aren't building the file from the passed
-    // string, but rather from the file pointed by it. Semantically, try_from_path is clearer
-    pub(crate) fn try_from_path(path: &str) -> FastExcelResult<Self> {
-        let sheets = open_workbook_auto(path)
-            .map_err(|err| FastExcelErrorKind::CalamineError(err).into())
-            .with_context(|| format!("Could not open workbook at {path}"))?;
-        let sheet_metadata = sheets.sheets_metadata().to_owned();
-        Ok(Self {
-            sheets: ExcelSheets::File(sheets),
-            sheet_metadata,
-            source: path.to_owned(),
-        })
-    }
-
-    #[cfg(feature = "python")]
     fn load_sheet_eager(
         data: &ExcelSheetData,
         pagination: Pagination,
@@ -175,35 +61,20 @@ impl ExcelReader {
         record_batch_from_data_and_columns(&available_columns, data, offset, limit)
     }
 
-    #[cfg(feature = "python")]
-    #[allow(clippy::too_many_arguments)]
     fn build_sheet<'py>(
         &mut self,
-        sheet_meta: CalamineSheet,
-        header_row: Option<usize>,
-        column_names: Option<Vec<String>>,
-        skip_rows: Option<usize>,
-        n_rows: Option<usize>,
-        schema_sample_rows: Option<usize>,
-        dtype_coercion: DTypeCoercion,
-        use_columns: Option<&Bound<'py, PyAny>>,
-        dtypes: Option<DTypes>,
+        idx_or_name: IdxOrName,
+        opts: LoadSheetOptions,
         eager: bool,
         py: Python<'py>,
     ) -> PyResult<Bound<'py, PyAny>> {
-        // calamine `header_row` is the first row of the range to be read.
-        // For us `header_row` can be `None` (meaning there is no header and we should start reading
-        // the data at the beginning)
-        let calamine_header_row = match (header_row, skip_rows) {
-            (None, None) | (Some(0), None) => HeaderRow::FirstNonEmptyRow,
-            (None, Some(_)) => HeaderRow::Row(0),
-            (Some(row), _) => HeaderRow::Row(row as u32),
-        };
-        // And our header row is simply the first row of the data if defined.
-        let data_header_row = header_row.and(Some(0));
+        let calamine_header_row = opts.calamine_header_row();
+        let data_header_row = opts.data_header_row();
 
-        let header = Header::new(data_header_row, column_names);
-        let selected_columns = Self::build_selected_columns(use_columns).into_pyresult()?;
+        let sheet_meta = self
+            .find_sheet_meta(idx_or_name)
+            .into_pyresult()?
+            .to_owned();
 
         if eager && self.sheets.supports_by_ref() {
             let range = self
@@ -211,21 +82,23 @@ impl ExcelReader {
                 .with_header_row(calamine_header_row)
                 .worksheet_range_ref(&sheet_meta.name)
                 .into_pyresult()?;
-            let pagination =
-                Pagination::new(skip_rows.unwrap_or(0), n_rows, &range).into_pyresult()?;
+            let pagination = Pagination::try_new(opts.skip_rows.unwrap_or(0), opts.n_rows, &range)
+                .into_pyresult()?;
+            let header = Header::new(data_header_row, opts.column_names);
             let rb = Self::load_sheet_eager(
                 &range.into(),
                 pagination,
                 header,
-                schema_sample_rows,
-                &selected_columns,
-                dtypes.as_ref(),
-                &dtype_coercion,
+                opts.schema_sample_rows,
+                &opts.selected_columns,
+                opts.dtypes.as_ref(),
+                &opts.dtype_coercion,
             )
             .into_pyresult()?;
 
             #[cfg(feature = "pyarrow")]
             {
+                use arrow_pyarrow::ToPyArrow;
                 rb.to_pyarrow(py).map(|py_object| py_object.into_bound(py))
             }
             #[cfg(not(feature = "pyarrow"))]
@@ -240,17 +113,17 @@ impl ExcelReader {
                 .with_header_row(calamine_header_row)
                 .worksheet_range(&sheet_meta.name)
                 .into_pyresult()?;
-            let pagination =
-                Pagination::new(skip_rows.unwrap_or(0), n_rows, &range).into_pyresult()?;
+            let pagination = opts.pagination(&range).into_pyresult()?;
+            let header = Header::new(data_header_row, opts.column_names);
             let sheet = ExcelSheet::try_new(
                 sheet_meta,
                 range.into(),
                 header,
                 pagination,
-                schema_sample_rows,
-                dtype_coercion,
-                selected_columns,
-                dtypes,
+                opts.schema_sample_rows,
+                opts.dtype_coercion,
+                opts.selected_columns,
+                opts.dtypes,
             )
             .into_pyresult()?;
 
@@ -271,7 +144,6 @@ impl ExcelReader {
         }
     }
 
-    #[cfg(feature = "python")]
     #[allow(clippy::too_many_arguments)]
     fn build_table<'py>(
         &mut self,
@@ -298,7 +170,7 @@ impl ExcelReader {
             }
         };
 
-        let pagination = Pagination::new(skip_rows, n_rows, table.data()).into_pyresult()?;
+        let pagination = Pagination::try_new(skip_rows, n_rows, table.data()).into_pyresult()?;
 
         let excel_table = ExcelTable::try_new(
             table,
@@ -328,24 +200,6 @@ impl ExcelReader {
     }
 }
 
-impl TryFrom<&[u8]> for ExcelReader {
-    type Error = FastExcelError;
-
-    fn try_from(bytes: &[u8]) -> Result<Self, Self::Error> {
-        let cursor = Cursor::new(bytes.to_vec());
-        let sheets = open_workbook_auto_from_rs(cursor)
-            .map_err(|err| FastExcelErrorKind::CalamineError(err).into())
-            .with_context(|| "Could not open workbook from bytes")?;
-        let sheet_metadata = sheets.sheets_metadata().to_owned();
-        Ok(Self {
-            sheets: ExcelSheets::Bytes(sheets),
-            sheet_metadata,
-            source: "bytes".to_owned(),
-        })
-    }
-}
-
-#[cfg(feature = "python")]
 #[pymethods]
 impl ExcelReader {
     pub fn __repr__(&self) -> String {
@@ -357,7 +211,7 @@ impl ExcelReader {
         self.sheets.table_names(sheet_name).into_pyresult()
     }
 
-    #[pyo3(signature = (
+    #[pyo3(name = "load_sheet", signature = (
         idx_or_name,
         *,
         header_row = 0,
@@ -371,7 +225,7 @@ impl ExcelReader {
         eager = false,
     ))]
     #[allow(clippy::too_many_arguments)]
-    pub(crate) fn load_sheet<'py>(
+    pub(crate) fn py_load_sheet<'py>(
         &mut self,
         idx_or_name: &Bound<'py, PyAny>,
         header_row: Option<usize>,
@@ -393,46 +247,20 @@ impl ExcelReader {
             .into())
             .into_pyresult();
         }
-        let sheet = idx_or_name
-            .try_into()
-            .and_then(|idx_or_name| match idx_or_name {
-                IdxOrName::Name(name) => {
-                    if let Some(sheet) = self.sheet_metadata.iter().find(|s| s.name == name) {
-                        Ok(sheet)
-                    } else {
-                        Err(FastExcelErrorKind::SheetNotFound(IdxOrName::Name(name.clone())).into()).with_context(||  {
-                            let available_sheets = self.sheet_metadata.iter().map(|s| format!("\"{}\"", s.name)).collect::<Vec<_>>().join(", ");
-                            format!(
-                                "Sheet \"{name}\" not found in file. Available sheets: {available_sheets}."
-                            )
-                        })
-                    }
-                }
-                IdxOrName::Idx(idx) => self
-                    .sheet_metadata
-                    .get(idx)
-                    .ok_or_else(|| FastExcelErrorKind::SheetNotFound(IdxOrName::Idx(idx)).into())
-                    .with_context(|| format!(
-                        "Sheet index {idx} is out of range. File has {} sheets.",
-                        self.sheet_metadata.len()
-                    )
-                ),
-            })
-            .into_pyresult()?.to_owned();
-
-        self.build_sheet(
-            sheet,
+        let idx_or_name = idx_or_name.try_into().into_pyresult()?;
+        let selected_columns = Self::build_selected_columns(use_columns).into_pyresult()?;
+        let opts = LoadSheetOptions {
             header_row,
             column_names,
             skip_rows,
             n_rows,
             schema_sample_rows,
             dtype_coercion,
-            use_columns,
+            selected_columns,
             dtypes,
-            eager,
-            py,
-        )
+        };
+
+        self.build_sheet(idx_or_name, opts, eager, py)
     }
 
     #[pyo3(signature = (
