+++ src/types/mod.rs
@@ -1,3 +1,8 @@
 pub(crate) mod dtype;
+pub(crate) mod excelreader;
+pub(crate) mod excelsheet;
+pub(crate) mod exceltable;
 pub(crate) mod idx_or_name;
-pub(crate) mod python;
+
+pub use excelreader::{ExcelReader, LoadSheetOptions};
+pub use excelsheet::ExcelSheet;
